Add files via upload
This code imports the Pandas and Matplotlib libraries in Python and performs some exploratory data analysis (EDA) on a drinks dataset. The dataset is read using Pandas, and various operations such as head(), tail(), index, data types, shape, mean, value counts, and filtering are performed on it to extract meaningful insights. The Matplotlib library is used to create visualizations such as bar plots and scatter plots to represent the data. The EDA aims to understand the alcohol consumption patterns across different continents and visualize them through plots.